# Conflicts:
#   client/.env
#   client/package.json
#   client/public/index.html
#   client/src/App.js
#   client/src/components/chat/Waiting.js
#   client/src/pages/Login.js
#   client/src/pages/Map.js

+++ server/src/main/java/B107/server/meerkat/dto/marker/MarkerDTO.java
@@ -1,6 +1,10 @@
 package B107.server.meerkat.dto.marker;
 
 import B107.server.meerkat.entity.Marker;
+import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
+import com.fasterxml.jackson.databind.annotation.JsonSerialize;
+import com.fasterxml.jackson.datatype.jsr310.deser.LocalDateTimeDeserializer;
+import com.fasterxml.jackson.datatype.jsr310.ser.LocalDateTimeSerializer;
 import lombok.*;
 
 import java.time.LocalDateTime;
@@ -9,15 +13,16 @@ import java.time.LocalDateTime;
 @Getter
 @NoArgsConstructor
 public class MarkerDTO {
-	private Float lat;
-	private Float lng;
+	private Double lat;
+	private Double lng;
 	private String location;
+
 	private LocalDateTime regDate; 		// 등록 시간
 	private LocalDateTime expDate; 		// 만료 시간
 
 
 	@Builder
-	public MarkerDTO(Float lat, Float lng, String location, LocalDateTime regDate, LocalDateTime expDate) {
+	public MarkerDTO(Double lat, Double lng, String location, LocalDateTime regDate, LocalDateTime expDate) {
 		this.lat = lat;
 		this.lng = lng;
 		this.location = location;
